LPS-66927 Move DDL related data handlers to the proper packages
